+++ src/main/func/rdb.ts
@@ -1,0 +1,1 @@
+// 추후 rdb 관리 기능 추가
