+++ src/main/index.ts
@@ -1,135 +1,9 @@
-import { app, BrowserWindow, shell, ipcMain, dialog } from "electron";
-import { release } from "node:os";
-import { join } from "node:path";
+import { BrowserWindow, ipcMain } from "electron"
+import { getOpenDirectoryDialog } from './func/electron'
 
-/**
- * The built directory structure
- *  ├─┬ dist-electron
- *  │ ├─┬ main
- *  │ │ └── index.js    > Electron-Main
- *  │ └─┬ preload
- *  │   └── index.js    > Preload-Scripts
- *  ├─┬ dist
- *  │ └── index.html    > Electron-Renderer
- */
-process.env.DIST_ELECTRON = join(__dirname, "..");
-process.env.DIST = join(process.env.DIST_ELECTRON, "../dist");
-process.env.VITE_PUBLIC = process.env.VITE_DEV_SERVER_URL
-  ? join(process.env.DIST_ELECTRON, "../public")
-  : process.env.DIST;
-
-// Disable GPU Acceleration for Windows 7
-if (release().startsWith("6.1")) app.disableHardwareAcceleration();
-
-// Set application name for Windows 10+ notifications
-if (process.platform === "win32") app.setAppUserModelId(app.getName());
-
-if (!app.requestSingleInstanceLock()) {
-  app.quit();
-  process.exit(0);
+const setMain = (window: BrowserWindow) => {
+  // 여기부터 api 처리 로직 작성
+  ipcMain.on("call:openFile", getOpenDirectoryDialog(window))
 }
 
-// Remove electron security warnings
-// This warning only shows in development mode
-// Read more on https://www.electronjs.org/docs/latest/tutorial/security
-// process.env['ELECTRON_DISABLE_SECURITY_WARNINGS'] = 'true'
-
-let win: BrowserWindow | null = null;
-// Here, you can also use other preload
-const preload = join(__dirname, "../preload/index.js");
-const url = process.env.VITE_DEV_SERVER_URL;
-const indexHtml = join(process.env.DIST, "index.html");
-
-async function createWindow() {
-  win = new BrowserWindow({
-    title: "Main window",
-    icon: join(process.env.VITE_PUBLIC, "favicon.ico"),
-    width: 400,
-    height: 670,
-    autoHideMenuBar: true,
-    webPreferences: {
-      preload,
-      // Warning: Enable nodeIntegration and disable contextIsolation is not secure in production
-      // Consider using contextBridge.exposeInMainWorld
-      // Read more on https://www.electronjs.org/docs/latest/tutorial/context-isolation
-      nodeIntegration: true,
-      contextIsolation: false,
-    },
-  });
-
-  if (process.env.VITE_DEV_SERVER_URL) {
-    // electron-vite-vue#298
-    win.loadURL(url);
-    // Open devTool if the app is not packaged
-    // win.webContents.openDevTools();
-  } else {
-    win.loadFile(indexHtml);
-  }
-
-  // Test actively push message to the Electron-Renderer
-  win.webContents.on("did-finish-load", () => {
-    win?.webContents.send("main-process-message", new Date().toLocaleString());
-  });
-
-  // Make all links open with the browser, not with the application
-  win.webContents.setWindowOpenHandler(({ url }) => {
-    if (url.startsWith("https:")) shell.openExternal(url);
-    return { action: "deny" };
-  });
-  // win.webContents.on('will-navigate', (event, url) => { }) #344
-}
-
-app.whenReady().then(createWindow);
-
-app.on("window-all-closed", () => {
-  win = null;
-  if (process.platform !== "darwin") app.quit();
-});
-
-app.on("second-instance", () => {
-  if (win) {
-    // Focus on the main window if the user tried to open another
-    if (win.isMinimized()) win.restore();
-    win.focus();
-  }
-});
-
-app.on("activate", () => {
-  const allWindows = BrowserWindow.getAllWindows();
-  if (allWindows.length) {
-    allWindows[0].focus();
-  } else {
-    createWindow();
-  }
-});
-
-// New window example arg: new windows url
-ipcMain.handle("open-win", (_, arg) => {
-  const childWindow = new BrowserWindow({
-    webPreferences: {
-      preload,
-      nodeIntegration: true,
-      contextIsolation: false,
-    },
-  });
-
-  if (process.env.VITE_DEV_SERVER_URL) {
-    childWindow.loadURL(`${url}#${arg}`);
-  } else {
-    childWindow.loadFile(indexHtml, { hash: arg });
-  }
-});
-
-// 여기부터 api 처리 로직 작성
-
-ipcMain.on("call:openFile", async (_event, _arg) => {
-  if (win === null) return;
-
-  const { canceled, filePaths } = await dialog.showOpenDialog(win, {
-    title: "Select Working Directory", // Dialog title
-    properties: ["openDirectory"], // Specify that it's an open file dialog
-  });
-
-  console.log(canceled, filePaths);
-  _event.reply("get:openFile", canceled ? "" : filePaths[0]);
-});
+export { setMain }
